refactor: extract getProcessedVideos into shared server function
Move video list transformation logic from the API route into
lib/videos.ts so it can be reused by both the API route and
server components.
